+++ packages/ui/tree-select/src/TreeSelect.tsx
@@ -1,11 +1,11 @@
-import React, { forwardRef, useCallback, useMemo, useState } from 'react'
+import React, { forwardRef, useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { cx, getPrefixCls } from '@hi-ui/classname'
 import { __DEV__ } from '@hi-ui/env'
 import { TreeSelectDataItem, TreeSelectAppearanceEnum } from './types'
 import { useUncontrolledToggle } from '@hi-ui/use-toggle'
 import { FlattedTreeNodeData, Tree, TreeNodeEventData } from '@hi-ui/tree'
 import { useUncontrolledState } from '@hi-ui/use-uncontrolled-state'
-import { Picker, PickerProps } from '@hi-ui/picker'
+import { Picker, PickerHelper, PickerProps } from '@hi-ui/picker'
 import { baseFlattenTree } from '@hi-ui/tree-utils'
 import { isArrayNonEmpty, isUndef } from '@hi-ui/type-assertion'
 import { uniqBy } from '@hi-ui/array-utils'
@@ -68,6 +68,7 @@ export const TreeSelect = forwardRef<HTMLDivElement | null, TreeSelectProps>(
       // popper,
       // ********* picker ********* //
       clearable,
+      onClear,
       invalid,
       displayRender: displayRenderProp,
       placeholder: placeholderProp,
@@ -85,6 +86,8 @@ export const TreeSelect = forwardRef<HTMLDivElement | null, TreeSelectProps>(
     ref
   ) => {
     const i18n = useLocaleContext()
+
+    const pickerInnerRef = useRef<PickerHelper>(null)
 
     const placeholder = isUndef(placeholderProp)
       ? i18n.get('treeSelect.placeholder')
@@ -238,9 +241,17 @@ export const TreeSelect = forwardRef<HTMLDivElement | null, TreeSelectProps>(
 
     const cls = cx(prefixCls, className)
 
+    useEffect(() => {
+      if (menuVisible) {
+        // 数据改变时更新弹窗显示位置，避免弹窗内容被遮挡
+        pickerInnerRef.current?.update()
+      }
+    }, [menuVisible, treeProps.expandedIds])
+
     return (
       <Picker
         ref={ref}
+        innerRef={pickerInnerRef}
         className={cls}
         {...rest}
         visible={menuVisible}
@@ -267,6 +278,7 @@ export const TreeSelect = forwardRef<HTMLDivElement | null, TreeSelectProps>(
               // disabled={disabled}
               size={size}
               clearable={clearable}
+              onClear={onClear}
               placeholder={placeholder}
               displayRender={displayRenderProp}
               prefix={prefix}
@@ -400,6 +412,10 @@ export interface TreeSelectProps
    * 是否可清空
    */
   clearable?: boolean
+  /**
+   * 点击关闭按钮时触发
+   */
+  onClear?: () => void
   /**
    * 设置展现形式
    */