+++ packages/ui/input/CHANGELOG.md
@@ -73,6 +73,11 @@
   - @hi-ui/env@5.0.0-canary.0
   - @hi-ui/func-utils@5.0.0-canary.0
   - @hi-ui/type-assertion@5.0.0-canary.0
+## 4.4.0
+
+### Minor Changes
+
+- [#3172](https://github.com/XiaoMi/hiui/pull/3172) [`a883d8e19`](https://github.com/XiaoMi/hiui/commit/a883d8e197446ef0e7cefcc5cc44d21ed1d0807f) Thanks [@zyprepare](https://github.com/zyprepare)! - feat(mock-input): 增加 onClear 参数 (#3171)
 
 ## 4.3.0