+++ packages/ui/picker/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @hi-ui/picker
+https://infra.mioffice.cn/hiui/# @hi-ui/picker
 
 ## 5.0.0-canary.7
 
@@ -98,6 +98,14 @@
   - @hi-ui/dom-utils@5.0.0-canary.0
   - @hi-ui/env@5.0.0-canary.0
   - @hi-ui/type-assertion@5.0.0-canary.0
+## 4.1.8
+
+### Patch Changes
+
+- [#3196](https://github.com/XiaoMi/hiui/pull/3196) [`80a909d2d`](https://github.com/XiaoMi/hiui/commit/80a909d2dae99d68d71f2ec6f4b210080d032ec0) Thanks [@KEH3](https://github.com/KEH3)! - fix(popper): ref 中增加更新位置方法解决上游组件弹窗被遮挡问题 (#3195)
+
+- Updated dependencies [[`80a909d2d`](https://github.com/XiaoMi/hiui/commit/80a909d2dae99d68d71f2ec6f4b210080d032ec0)]:
+  - @hi-ui/popper@4.1.7
 
 ## 4.1.7